feat: add snapcraft packaging
This commit adds snapcraft packaging of the Ubuntu Packaging Guide. The
snap provides an HTTP file server that hosts the Ubuntu Packaging Guide
locally. This allows users to access the guide even when they are
offline or have a poor internet connection.

The snap build for amd64 and arm64.

See: https://snapcraft.io/ubuntu-packaging-guide