refactor: move category index files to category dir
This has the benefit that the index file are grouped with the
content, which makes it easier to find them. I noticed that I very
often spend a few seconds searching for them up to the amount to
justify this change.

note: I will fix the ref paths to and from the index files in the next
commit to keep the file history. If I would do it in one go, git would
think that I deleted files and created completely new ones.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>