feat: add Ubuntu package archive overview diagram
This commit adds a mermaid diagram that gives a
quick overview over the structure and terminology
of the Ubuntu package archive.

note: This diagram is a copy from the Ubuntu server
      documentation. See the Origin footer below.

This commit also adds a sphinx mermaid extension
to render the mermaid diagram.

todo: The diagram has a low contrast in dark mode.

Origin: https://documentation.ubuntu.com/server/tutorial/managing-software/#where-do-packages-come-from
Fixes: #82